move list components and functions to its own
